# fix(perf): remove DOM mutation overhead from fog-of-war prefetch