--- a/contributors.yml
+++ b/contributors.yml
@@ -75,6 +75,7 @@
 - david-crespo
 - dcblair
 - decadentsavant
+- developit
 - dgrijuela
 - DigitalNaut
 - dmitrytarassov